% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_predictor.R
\name{predict.TextPredictor}
\alias{predict.TextPredictor}
\title{Predict Next Word
Using the Text Predictor and tokens as input, predict a table of the next word}
\usage{
\method{predict}{TextPredictor}(object, tokens)
}
\arguments{
\item{object}{The TextPredictor object}

\item{tokens}{A character vector of tokens representing word history}
}
\value{
A data.table of next best words in order
}
\description{
Predict Next Word
Using the Text Predictor and tokens as input, predict a table of the next word
}
\examples{
predict(mytextpred, c("hello","my","darling"))
}

