% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_predictor.R
\name{tokenize_from_input}
\alias{tokenize_from_input}
\title{Tokenize from English Text Input
Returns tokens from an English text input for last sentence only. Not flexible - to be used for next word prediction
only.}
\usage{
tokenize_from_input(string)
}
\arguments{
\item{string}{Some English text}
}
\description{
Tokenize from English Text Input
Returns tokens from an English text input for last sentence only. Not flexible - to be used for next word prediction
only.
}
\examples{
tokenize_from_input("hello my darling")
}

