% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_predictor.R
\name{predict.TextPredictor}
\alias{predict.TextPredictor}
\title{Predict Next Word
Using the Text Predictor and tokens as input, predict a table of the next word}
\usage{
\method{predict}{TextPredictor}(object, string)
}
\arguments{
\item{object}{The TextPredictor object}

\item{string}{An English language string of words}
}
\value{
A data.table of next best words in order
}
\description{
Predict Next Word
Using the Text Predictor and tokens as input, predict a table of the next word
}
\examples{
predict(mytextpred, "hello my darling"))
}

